languagemodel/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-textanalytics==5.2.0
azure-storage-file-datalake
azure-identity
python-dotenv
azure-storage-blob
PyPDF2
python-docx
pandas
